- Creating the refinery. - Fix partially broken tank texture.

+++ com/projectzed/mod/renderer/FluidTankRenderer.java
@@ -6,6 +6,11 @@
 */
 package com.projectzed.mod.renderer;
 
+import com.hockeyhurd.api.math.Vector3;
+import com.hockeyhurd.api.util.TessellatorHelper;
+import com.projectzed.mod.tileentity.container.TileEntityFluidTankBase;
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.client.renderer.OpenGlHelper;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.client.renderer.texture.TextureMap;
@@ -14,15 +19,7 @@ import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.IIcon;
 import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.fluids.FluidStack;
-
 import org.lwjgl.opengl.GL11;
-
-import com.hockeyhurd.api.math.Vector3;
-import com.hockeyhurd.api.util.TessellatorHelper;
-import com.projectzed.mod.tileentity.container.TileEntityFluidTankBase;
-
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
 
 /**
  * Class handling all client side rendering code for fluid tanks.
@@ -98,7 +95,7 @@ public class FluidTankRenderer extends TileEntitySpecialRenderer {
 		Tessellator tess = tessHelp.tess;
 		tess.startDrawingQuads();
 
-		byte counter = 0;
+		int counter = 0;
 		for (byte valve : te.getSidedArray()) {
 
 			float max = 1f;
